koya-pay/backend/requirements.txt
Flask==2.1.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
opencv-python==4.5.3.20210927
numpy==1.21.2
Pillow==8.4.0
face_recognition==1.3.0
requests==2.26.0